Add copyrights and replace dupe test model
