+++ python/openai/openai/tests/test_models/identity_py/config.pbtxt
@@ -23,38 +23,29 @@
 # OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
 backend: "python"
-
-max_batch_size: 0
-
-model_transaction_policy {
-  decoupled: True
-}
+max_batch_size: 64
 
 input [
   {
-    name: "text_input"
-    data_type: TYPE_STRING
-    dims: [ 1, 1 ]
-  },
-  {
-    name: "stream"
-    data_type: TYPE_BOOL
-    dims: [ 1, 1 ]
+    name: "INPUT0"
+    data_type: TYPE_FP32
+    dims: [ -1 ]
   }
 ]
 
 output [
   {
-    name: "text_output"
-    data_type: TYPE_STRING
-    dims: [ 1, -1 ]
+    name: "OUTPUT0"
+    data_type: TYPE_FP32
+    dims: [ -1 ]
   }
 ]
 
 instance_group [
   {
     count: 1
-    kind: KIND_MODEL
+    kind : KIND_CPU
   }
 ]
